+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### New Features
 
 - Show voting results of live vote even the participant has not yet voted or saved ([#1425](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1425))
+- Add mobile drawer ([#1120](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1120))
 
 ### Improvements to the user experience